IG2LLM/requirements.txt
openai
dotenv
tabulate
tqdm
langchain
pydantic
google-genai
anthropic
numpy